Update: Button path common으로 변경
