--- src/components/Regist/atom/Button/style.ts
@@ -1,14 +1,0 @@
-import styled from '@emotion/styled';
-import { Palette } from 'style/global';
-
-export const Wrapper = styled.button`
-  width: 320px;
-  padding: 0.825rem 7rem;
-  background: ${Palette.PRIMARY_P5};
-  border: none;
-  border-radius: 10px;
-  color: #fff;
-  font-size: 20px;
-  font-weight: 500;
-  cursor: pointer;
-`;
