+++ src/components/Regist/organisms/index.tsx
@@ -1,4 +1,4 @@
-import Button from '../atom/Button';
+import Button from '../../common/atoms/Button';
 import BannerBox from '../molecules/BannerBox';
 import InputBox from '../molecules/InputBox';
 import * as S from './style';
@@ -9,7 +9,9 @@ const Regist = () => {
       <p>등록하기</p>
       <BannerBox />
       <InputBox />
-      <Button text="등록하기" />
+      <S.BtnBox>
+        <Button text="등록하기" fontsize="20" />
+      </S.BtnBox>
     </S.Wrapper>
   );
 };
